% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Crosstab.Correlations}
\alias{Crosstab.Correlations}
\title{Creates a crosstab in Excel and Word of correlations between opinions and changes in opinion.}
\usage{
Crosstab.Correlations(
  Dataset,
  Template,
  Outputs,
  Question_Beg,
  Question_End,
  Demographic_Beg,
  Demographic_End
)
}
\description{
Creates a crosstab in Excel and Word of correlations between opinions and changes in opinion.
}
