% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Crosstab.Demographics}
\alias{Crosstab.Demographics}
\title{Creates a crosstab in Excel and Word comparing demographics}
\usage{
Crosstab.Demographics(
  Dataset,
  Template,
  Group1,
  Group2,
  Outputs,
  Export,
  Alpha,
  Only_Significant
)
}
\description{
Creates a crosstab in Excel and Word comparing demographics
}
