% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Crosstab.Nominal}
\alias{Crosstab.Nominal}
\title{Creates a crosstab in Excel and Word comparing opinions.}
\usage{
Crosstab.Nominal(
  Dataset,
  Templates,
  Outputs,
  Group1,
  Group2,
  Report_Demographic,
  Alpha,
  Only_Significant,
  Only_Means,
  API_Key
)
}
\description{
Creates a crosstab in Excel and Word comparing opinions.
}
