% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Crosstab.Opinions.Detailed}
\alias{Crosstab.Opinions.Detailed}
\title{Creates a crosstab in Excel and Word comparing opinions between two demographic categories.}
\usage{
Crosstab.Opinions.Detailed(
  Dataset,
  Template,
  Group1,
  Group2,
  Demographics,
  Outputs,
  Export,
  Alpha,
  Only_Means,
  Only_Significant
)
}
\description{
Creates a crosstab in Excel and Word comparing opinions between two demographic categories.
}
