% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Crosstab.Opinions}
\alias{Crosstab.Opinions}
\title{Creates a crosstab in Excel and Word comparing opinions.}
\usage{
Crosstab.Opinions(
  Dataset,
  Template,
  Group1,
  Group2,
  Outputs,
  Export,
  Alpha,
  Only_Significant,
  Only_Means
)
}
\description{
Creates a crosstab in Excel and Word comparing opinions.
}
