% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Crosstab.Ordinal}
\alias{Crosstab.Ordinal}
\title{Creates a crosstab in Excel and Word comparing demographics}
\usage{
Crosstab.Ordinal(
  Dataset,
  Templates,
  Outputs,
  Group1,
  Group2,
  Alpha,
  Only_Significant
)
}
\description{
Creates a crosstab in Excel and Word comparing demographics
}
