% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Dataset.Organize}
\alias{Dataset.Organize}
\title{Strips the dataset of all columns not present in the codebook.}
\usage{
Dataset.Organize(Codebook, Datasets)
}
\description{
Strips the dataset of all columns not present in the codebook.
}
