% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Dataset}
\alias{Dataset}
\title{Returns list of data frames.
Returns list of data frames.}
\usage{
Dataset(Codebook, Datasets)
}
\arguments{
\item{Codebook}{is a sheet containing a correctly formatted codebook for the dataset file.}

\item{Datasets}{are the unformatted datasets to be formatted to match the codebook.}
}
\description{
This function returns a list of data frames and the codebook formatted for use in the "Results" function.
}
