% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Export_Report}
\alias{Export_Report}
\title{Exports a report on crosstab in Word.}
\usage{
Export_Report(Outputs, Crosstabs, File_Name, Name_Group, Demographic_Category)
}
\description{
Exports a report on crosstab in Word.
}
