% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Export_Word}
\alias{Export_Word}
\title{Exports crosstab in Word.}
\usage{
Export_Word(
  Export,
  Outputs,
  Template,
  Crosstabs,
  Codebook,
  File_Name,
  Document_Title,
  Type,
  Demographic_Category
)
}
\description{
Exports crosstab in Word.
}
