% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Multiple.Regression}
\alias{Multiple.Regression}
\title{Creates a crosstab of regressions.}
\usage{
Multiple.Regression(
  Dataset,
  Template,
  Model,
  Group1,
  Group2,
  Outputs,
  Export,
  Alpha,
  Only_Significant
)
}
\description{
Creates a crosstab of regressions.
}
