% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Ordinal}
\alias{Ordinal}
\title{Creates a crosstab in Excel and Word comparing demographics}
\usage{
Ordinal(Dataset, Template, Outputs, Group1, Group2, Alpha, Only_Significant)
}
\description{
Creates a crosstab in Excel and Word comparing demographics
}
