% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Results}
\alias{Results}
\title{Creates crosstabs and a report on results.}
\usage{
Results(
  Dataset,
  Templates,
  Outputs,
  Group1,
  Group2,
  Report_Demographic,
  Alpha,
  Only_Means,
  Only_Significant,
  API_Key
)
}
\description{
Creates crosstabs and a report on results.
}
