% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Test_Chi_Squared}
\alias{Test_Chi_Squared}
\title{Defines a function that runs a chi-squared test.}
\usage{
Test_Chi_Squared(ValuetoMark, Group1, Group2)
}
\description{
Defines a function that runs a chi-squared test.
}
