% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Dataset.Organize}
\alias{Dataset.Organize}
\title{Strips a dataset of all  for crosstab creation.}
\usage{
Dataset.Organize(Codebook, Datasets)
}
\description{
Strips a dataset of all  for crosstab creation.
}
