% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Export_Word}
\alias{Export_Word}
\title{Exports a crosstab in Word.}
\usage{
Export_Word(
  Crosstabs,
  Outputs,
  File_Name,
  Name_Group,
  Template,
  Document_Title,
  Type,
  Demographic_Category,
  Codebook
)
}
\description{
Exports a crosstab in Word.
}
